requirements.txt
# Core dependencies
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
pydantic-settings>=2.0.0
python-multipart>=0.0.6

# Neural/ML
torch>=2.2.0
transformers>=4.35.0
numpy>=1.24.3
scipy>=1.11.4
faiss-cpu>=1.7.4  # Use faiss-gpu for GPU support

# Database
asyncpg>=0.29.0
sqlalchemy>=2.0.23
alembic>=1.12.1
redis>=5.0.1

# API
strawberry-graphql[fastapi]>=0.215.1
websockets>=12.0
pyjwt>=2.10.1

# Utilities
python-dotenv>=1.0.0
pyyaml>=6.0.1
click>=8.1.7

# Development
pytest>=7.4.3
pytest-asyncio>=0.21.1
black>=23.11.0
flake8>=6.1.0
mypy>=1.7.0
isort>=5.12.0

# Documentation
sphinx>=7.2.6
sphinx-rtd-theme>=1.3.0
